refactor(core): split gesture shared state from tap/doubletap logic
Extract per-element pointer listeners, binding registry, and matching
into shared.ts. gesture.ts imports TapRecognizer and wires the handler.
A future createSwipeGesture in its own file would import only shared.ts
without pulling in tap/doubletap code.
